Do final Google refactoring
