+++ src/main/java/it/polpetta/libris/image/contract/IAbstractImageFactoryReverseSearchProvider.java
@@ -1,0 +1,9 @@
+package it.polpetta.libris.image.contract;
+
+import it.polpetta.libris.contract.IAbstractFactoryReverseSearchProvider;
+
+/**
+ * Created by dpolonio on 10/05/17.
+ */
+public interface IAbstractImageFactoryReverseSearchProvider extends IAbstractFactoryReverseSearchProvider {
+}
