+++ src/main/java/it/polpetta/libris/image/google/contract/IAbstractGoogleImageFactoryReverseSearchProvider.java
@@ -1,0 +1,11 @@
+package it.polpetta.libris.image.google.contract;
+
+import it.polpetta.libris.image.contract.IAbstractImageFactoryReverseSearchProvider;
+
+/**
+ * Created by dpolonio on 10/05/17.
+ */
+public interface IAbstractGoogleImageFactoryReverseSearchProvider extends IAbstractImageFactoryReverseSearchProvider {
+
+
+}
